Add wiki selection into the wiki layer and remove unused wiki-selection component
