+++ app/components/layer-file/style.scss
@@ -5,9 +5,3 @@
 .file-browser .file-browser-list {
     height: 300px;
 }
-
-.select-file {
-    border: 2px dashed #ccc;
-    text-align: center;
-    cursor: pointer;
-}
