+++ app/components/layer-file/template.hbs
@@ -15,7 +15,7 @@
                 {{file-browser rootItem=node openFile=(action 'fileDetail') }}
             </div>
         {{else}}
-            <div class="select-file p-v-md" {{action 'showSelect'}}>
+            <div class="select-box p-v-md" {{action 'showSelect'}}>
                 <p>
                     {{#if layer.settings.values.downloadLink}}
                         Change selection
