+++ app/components/layer-wiki/component.js
@@ -5,6 +5,7 @@ import config from 'ember-get-config';
 export default Ember.Component.extend({
     session: Ember.inject.service(),
     currentWiki: null,
+    showSelect: false,
     reloadWiki: Ember.computed('layer.settings.wikiId', function(){
         this.get('node.wikis').then((result)=>{
             if(!this.get('layers.settings.wikiID')){
@@ -44,8 +45,34 @@ export default Ember.Component.extend({
             this.get('layer').set('wikiContent', data);
         });
     },
+    wikis: Ember.computed('node', function(){
+        let wikis = this.get('node.wikis');
+        let wikiId = this.get('layer.settings.wikiId');
+        this.get('node.wikis').then((result)=>{
+            let loadedWiki = result.find(function (item) {
+                return item.get('id') === wikiId;
+            });
+            this.set('currentWiki', loadedWiki);
+        });
+        return wikis;
+    }),
     init(){
         this._super(...arguments);
         this.get('reloadWiki');
+    },
+    actions: {
+        showSelect(){
+            this.set('showSelect', true);
+        },
+        hideSelect(){
+            this.set('showSelect', false);
+        },
+        selectWiki(selected){
+            this.set('layer.settings.wikiId', selected.get('id'));
+            this.set('currentWiki', selected);
+            this.loadCurrentWiki();
+            this.set('showSelect', false);
+
+        }
     }
 });
