--- app/components/wiki-select/template.hbs
@@ -1,6 +1,0 @@
-{{#power-select
-    options=wikis
-    selected=currentWiki
-    onchange=(action "selectWiki") as |wiki|}}
-    {{wiki.name}}
-{{/power-select}}
