+++ app/styles/app.scss
@@ -23,6 +23,13 @@ body {
     background: #eee;
 }
 
+.select-box {
+    border: 2px dashed #ccc;
+    text-align: center;
+    cursor: pointer;
+}
+
+
 /* Pages-nav */
 .pages-nav ul{
     list-style-type: none;
