--- tests/integration/components/wiki-select/component-test.js
@@ -1,25 +1,0 @@
-import { moduleForComponent, test } from 'ember-qunit';
-import hbs from 'htmlbars-inline-precompile';
-
-moduleForComponent('wiki-select', 'Integration | Component | wiki select', {
-  integration: true
-});
-
-test('it renders', function(assert) {
-
-  // Set any properties with this.set('myProperty', 'value');
-  // Handle any actions with this.on('myAction', function(val) { ... });
-
-  this.render(hbs`{{wiki-select}}`);
-
-  assert.equal(this.$().text().trim(), '');
-
-  // Template block usage:
-  this.render(hbs`
-    {{#wiki-select}}
-      template block text
-    {{/wiki-select}}
-  `);
-
-  assert.equal(this.$().text().trim(), 'template block text');
-});
